# Conflicts:
#	src/algorithms/rhythm/tempotapdegara.h
#	src/algorithms/rhythm/tempotapmaxagreement.h
#	src/algorithms/stats/poolaggregator.cpp

+++ test/src/unittest/rhythm/test_onsetrate.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (C) 2006-2013  Music Technology Group - Universitat Pompeu Fabra
+# Copyright (C) 2006-2016  Music Technology Group - Universitat Pompeu Fabra
 #
 # This file is part of Essentia
 #
@@ -32,7 +32,7 @@ class TestOnsetRate(TestCase):
 
     def testZero(self):
         # zeros should return no onsets
-        size = 5*sr
+        size = int(5*sr)
         times, rate = OnsetRate()(zeros(size))
         self.assertEqualVector(times, [])
         self.assertEqual(rate, 0)
